requirements.txt
Flask
pydantic
Flask-Pydantic
psycopg2-binary
